Move linear code prime admission to operations
